Add publication comments to `UseCreateComment` example
