+++ examples/web-wagmi/src/publications/UseCreateComment.tsx
@@ -1,84 +1,42 @@
-import {
-  CollectPolicyType,
-  ContentFocus,
-  ProfileFieldsFragment,
-  ReferencePolicy,
-  useCreateComment,
-} from '@lens-protocol/react';
 import { useState } from 'react';
 
 import { LoginButton } from '../components/auth/LoginButton';
 import { WhenLoggedIn, WhenLoggedOut } from '../components/auth/auth';
-import { upload } from '../upload';
-
-type CreatePostProps = {
-  activeProfile: ProfileFieldsFragment;
-};
-
-export function CreateComment({ activeProfile }: CreatePostProps) {
-  const [publicationId, setPublicationId] = useState<string>('0x1b-0x0118');
-  const [content, setContent] = useState<string>('');
-
-  const { create, error, isPending } = useCreateComment({ profile: activeProfile, upload });
-
-  const submit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (content.length === 0) {
-      return;
-    }
-    await create({
-      publicationId,
-      content,
-      profileId: activeProfile.id,
-      contentFocus: ContentFocus.TEXT,
-      locale: 'en',
-      collect: {
-        type: CollectPolicyType.NO_COLLECT,
-      },
-      reference: ReferencePolicy.ANYBODY,
-    });
-
-    setContent('');
-  };
-
-  return (
-    <form onSubmit={submit}>
-      {error && <p>{error.message}</p>}
-
-      <p>
-        <label htmlFor="publicationId">Publication id</label>
-        <input
-          id="publicationId"
-          type="text"
-          disabled={isPending}
-          value={publicationId}
-          onChange={(event) => setPublicationId(event.target.value)}
-        />
-      </p>
-
-      <textarea
-        rows={3}
-        placeholder="Say gm...?"
-        style={{ resize: 'none' }}
-        disabled={isPending}
-        onChange={(event) => setContent(event.target.value)}
-        value={content}
-      ></textarea>
-
-      <button type="submit" disabled={isPending}>
-        Comment
-      </button>
-    </form>
-  );
-}
+import { PublicationComments } from './components/PublicationComments';
+import { CommentComposer } from './components/CommentComposer';
 
 export function UseCreateComment() {
+  const [publicationId, setPublicationId] = useState<string>('0x1b-0x0118');
+
   return (
     <div>
       <h1>
         <code>useCreateComment</code>
       </h1>
-      <WhenLoggedIn>{({ profile }) => <CreateComment activeProfile={profile} />}</WhenLoggedIn>
+      <WhenLoggedIn>
+        {({ profile }) => (
+          <>
+            <p>
+              <label htmlFor="publicationId">Publication id</label>
+              <input
+                id="publicationId"
+                type="text"
+                value={publicationId}
+                onChange={(event) => setPublicationId(event.target.value)}
+              />
+            </p>
+
+            {publicationId && (
+              <>
+                <CommentComposer activeProfile={profile} publicationId={publicationId} />
+
+                <p>Publication comments:</p>
+                <PublicationComments publicationId={publicationId} />
+              </>
+            )}
+          </>
+        )}
+      </WhenLoggedIn>
       <WhenLoggedOut>
         <div>
           <p>Log in to create a post.</p>
